requirements.txt
# Pretty Good AI — Voice Patient Bot
# Python 3.9+

# Telephony & WebSocket
twilio>=8.0.0
flask>=3.0.0
flask-sockets>=0.2.1
gevent>=23.0.0
gevent-websocket>=0.10.1

# Speech: STT & TTS
openai>=1.0.0

# Audio: mulaw 8kHz for Twilio
pydub>=0.25.1
numpy>=1.24.0

# Config & CLI
python-dotenv>=1.0.0
requests>=2.28.0

# Optional: faster STT (uncomment if you prefer)
# deepgram-sdk>=3.0.0
